refactor: Started work on Job class
Most of the job handlers are being refactored
to work with individual jobs as a function within
the class itself.

Then we can simply check the class variables in a loop
just once when we want to iterate through and prompt user for action.
This should keep things tidy, be faster and make everything easier.